dk finished? (REMOVE tokens...)
