+++ app/src/main/java/com/homework/todolist/ui/screen/about/DivNavigationHandler.kt
@@ -1,0 +1,8 @@
+package com.homework.todolist.ui.screen.about
+
+/**
+ * Handler for custom Div navigation action
+ */
+internal interface DivNavigationHandler {
+    fun dismiss()
+}
