+++ build-logic/src/main/kotlin/android-app-convention.gradle.kts
@@ -5,23 +5,6 @@ plugins {
     id("org.jetbrains.kotlin.android")
 }
 
-configure<BaseAppModuleExtension> {
-    baseAndroidConfig()
-    buildFeatures {
-        compose = true
-        buildConfig = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = "1.5.8"
-    }
-
-    defaultConfig {
-        val baseUrl = providers.environmentVariable("BASE_URL")
-        buildConfigField("String", "BASE_URL", "\"${baseUrl.get()}\"")
-        manifestPlaceholders["YANDEX_CLIENT_ID"] = providers.environmentVariable("CLIENT_ID")
-    }
-}
-
 dependencies {
     implementation(libs.core.ktx)
     implementation(libs.lifecycle.runtime.ktx)
@@ -33,3 +16,22 @@ dependencies {
     implementation(libs.material3)
     testImplementation(libs.junit)
 }
+
+configure<BaseAppModuleExtension> {
+    baseAndroidConfig()
+    buildFeatures {
+        compose = true
+        buildConfig = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = "1.5.8"
+    }
+
+    defaultConfig {
+        val baseUrl = "\"https://hive.mrdekk.ru/todo/\"" // providers.environmentVariable("BASE_URL")
+        buildConfigField("String", "BASE_URL", baseUrl.toString())
+        val clientId = "f9466f41f83c4266b6b55d1d67649709" // providers.environmentVariable("CLIENT_ID")
+        manifestPlaceholders["YANDEX_CLIENT_ID"] = clientId.toString()
+    }
+}
+
