+++ build-logic/src/main/kotlin/com/homework/todolist/telegramTask/TelegramReporterTask.kt
@@ -1,0 +1,44 @@
+package com.homework.todolist.telegramTask
+
+import kotlinx.coroutines.runBlocking
+import org.gradle.api.DefaultTask
+import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.provider.Property
+import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.InputDirectory
+import org.gradle.api.tasks.TaskAction
+import javax.inject.Inject
+
+abstract class TelegramReporterTask @Inject constructor(
+    private val telegramApi: TelegramApi
+) : DefaultTask() {
+
+    @get:InputDirectory
+    abstract val apkDir: DirectoryProperty
+
+    @get:Input
+    abstract val token: Property<String>
+
+    @get:Input
+    abstract val chatId: Property<String>
+
+    @TaskAction
+    fun report() {
+        val token = token.get()
+        val chatId = chatId.get()
+        apkDir.get().asFile.listFiles()
+            ?.filter { it.name.endsWith(".apk") }
+            ?.forEach {
+                runBlocking {
+                    telegramApi.sendMessage("Build finished", token, chatId).apply {
+                        println("Status = $status")
+                    }
+                }
+                runBlocking {
+                    telegramApi.upload(it, token, chatId).apply {
+                        println("Status = $status")
+                    }
+                }
+            }
+    }
+}
